Delete the dead `importado` package and rename the applier fixture
Event import was removed by the cutover, leaving the `importado` demo package
(`Dot`/`DotSpace`/`MoveMessages` + protos) orphaned with no references — deleted.

Renamed `IndecisiveEngineAggregate` to `EngineAggregateWithApplier`: its original
"two appliers are ambiguous" premise is obsolete now that any `@Apply` applier is
rejected, so it reads as the canonical single-applier "reject appliers" fixture.
