--- server/src/testFixtures/java/io/spine/server/aggregate/given/klasse/IndecisiveEngineAggregate.java
@@ -1,55 +1,0 @@
-/*
- * Copyright 2026, TeamDev. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.server.aggregate.given.klasse;
-
-import io.spine.server.aggregate.Aggregate;
-import io.spine.server.aggregate.Apply;
-import io.spine.server.aggregate.given.klasse.command.StartEngine;
-import io.spine.server.aggregate.given.klasse.event.EngineStarted;
-import io.spine.server.command.Assign;
-
-/**
- * A test-only aggregate that still declares an {@link Apply @Apply} event applier.
- *
- * <p>Event sourcing has been removed, so appliers are no longer supported. Building the model
- * for this class fails fast with a {@link io.spine.server.model.ModelError ModelError}.
- */
-public class IndecisiveEngineAggregate extends Aggregate<EngineId, Engine, Engine.Builder> {
-
-    @Assign
-    EngineStarted handle(StartEngine command) {
-        var id = command.getId();
-        return EngineStarted.newBuilder()
-                .setId(id)
-                .build();
-    }
-
-    @Apply
-    private void on(EngineStarted event) {
-        // Declaring any `@Apply` applier makes model building fail with a `ModelError`.
-    }
-}
