--- server/src/testFixtures/proto/spine/test/aggregate/importado/events.proto
@@ -1,51 +1,0 @@
-/*
- * Copyright 2022, TeamDev. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-syntax = "proto3";
-
-package spine.test.aggregate.command;
-
-import "spine/options.proto";
-
-option (type_url_prefix) = "type.spine.io";
-option java_package="io.spine.server.aggregate.given.importado.event";
-option java_outer_classname = "MovementEventsProto";
-option java_multiple_files = true;
-
-import "spine/test/aggregate/importado/movement.proto";
-
-// An object has moved.
-message Moved {
-
-    // An identifier of the object that moved.
-    ObjectId object = 1 [(required) = true];
-
-    // The direction in which the object moved.
-    Direction direction = 2 [(required) = true];
-
-    // A position of the object after the move.
-    // This field is optional to allow objects disclosing at their will.
-    Point current_position = 3;
-}
